refactor: consolidate to one crate + passing tests
