+++ Cargo.toml
@@ -1,10 +1,44 @@
-[workspace]
-resolver = "2"
-members = [
-  "ekm/client",
-  "ekm/server",
-  "ekm/types",
-  "ekm/providers/google_confidential_space/client",
-  "ekm/providers/google_confidential_space/server",
-  "ekm/providers/google_confidential_space/types",
-]
+[package]
+name = "caratls"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+tokio = { version = "1.42", default-features = false, features = ["io-util"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_cbor = "0.11"
+thiserror = "2"
+rcgen = "0.13"
+rustls = { version = "0.23" }
+tokio-rustls = { version = "0.26" }
+
+# Not WASM, not iOS
+[target.'cfg(not(any(target_arch = "wasm32", target_os = "ios", target_os = "ios-sim")))'.dependencies]
+rustls-pki-types = { version = "1.10" }
+rustls = { version = "0.23" }
+tokio-rustls = { version = "0.26" }
+
+# WASM or iOS
+[target.'cfg(any(target_arch = "wasm32", target_os = "ios", target_os = "ios-sim"))'.dependencies]
+rustls = { version = "0.23", default-features = false, features = ["ring"] }
+tokio-rustls = { version = "0.26", default-features = false, features = [
+  "ring",
+] }
+
+# WASM
+[target.'cfg(any(target_arch = "wasm32"))'.dependencies]
+rustls-pki-types = { version = "1.10", features = ["web"] }
+
+# iOS
+[target.'cfg(any(target_os = "ios", target_os = "ios-sim"))'.dependencies]
+rustls-pki-types = { version = "1.10" }
+
+
+[dev-dependencies]
+tokio = { version = "1.42", features = [
+  "net",
+  "io-util",
+  "rt",
+  "macros",
+  "time",
+] }
