--- ekm/providers/google_confidential_space/server/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "caratls_ekm_google_confidential_space_server"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-tokio = { version = "1.42", default-features = false, features = ["io-util"] }
-ekm_gcs_types = { package = "caratls_ekm_google_confidential_space_types", path = "../types" }
-ekm_server = { package = "caratls_ekm_server", path = "../../../server" }
-hyper = { version = "1.5", features = ["client", "http1"] }
-http-body-util = { version = "0.1" }
-hyper-util = { version = "0.1", features = ["tokio"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = { version = "1.0" }
-hex = { version = "0.4" }
-jsonwebtoken = { version = "9.3" }
-reqwest = { version = "0.12" }
