--- ekm/types/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "caratls_ekm_types"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
