+++ src/providers/google_confidential_space/client.rs
