+++ src/providers/google_confidential_space/types.rs
