+++ src/providers/lib.rs
@@ -1,0 +1,1 @@
+pub mod google_confidential_space;
